app/requirements.txt
pandas
PySide6 
numpy 
watchdog
psycopg2-binary
openpyxl